test: Fix and update tests after create UserPolicy

+++ resources/views/livewire/pages/users/index.blade.php
@@ -41,6 +41,12 @@ new class extends Component {
     {
         $this->authorize('delete', $user);
 
+        if (auth()->id() === $user->id) {
+            $this->redirectRoute('settings.profile');
+
+            return;
+        }
+
         if ($user->avatar) {
             $path = str($user->avatar)->after('/storage/');
             \Storage::disk('public')->delete($path);
